startBlock device list as select tag.
It turns out Android Chrome hijacks the select tag and does a full
screen UI for choosing items, This is a ridiculous carryover from the
small screen phone world, that ignores all current CSS settings.
However, the implementation is simple  and works very well on FF,
Chrome, Safari desktop. So Im checking it in in case it can be used in
the future. Bummer.